pip_requirements.txt
biopython~=1.78
matplotlib~=3.9.2
nglview~=3.1.4
numpy~=1.24.3
pandas~=2.2.2
plotly~=5.24.1
py3dmol~=2.4.2
requests~=2.32.3
scikit-learn~=1.5.1
seaborn~=0.13.2
statsmodels~=0.14.2
rdkit~=2024.09.3
requests~=2.32.3
torch~=2.3.0
tqdm~=4.66.5
transformers~=4.45.2
umap-learn~=0.5.4
igraph~=0.11.8
plotly~=5.24.1